stubbing

+++ fei.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.21;
 
 struct Fei {
     bytes Mu;
-    uint64 Context;
-    uint64 Operation;
-    uint64 Entity;
+    uint256 Context;
+    uint256 Operation;
+    uint256 Entity;
 }
